+++ imagepy/__init__.py
@@ -1,0 +1,5 @@
+__version__ = "1.0.0"
+
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).parent
