--- src/lab4/filters.py
@@ -1,34 +1,0 @@
-blur_filters = [
-    [[1, 1, 1], [1, 1, 1], [1, 1, 1]],
-    [[1, 1, 1], [1, 8, 1], [1, 1, 1]],
-    [[1, 2, 1], [2, 4, 2], [1, 2, 1]],
-]
-
-sharpen_filters = [
-    [[0, -1, 0], [-1, 4, -1], [0, -1, 0]],
-    [[-1, -1, -1], [-1, 8, -1], [-1, -1, -1]],
-    [[1, -2, 1], [-2, 4, -2], [1, -2, 1]],
-]
-
-edge_detection_filters = {
-    "E": [[-1, 0, 1], [-2, 0, 2], [-1, 0, 1]],
-    "SE": [[-2, -1, 0], [-1, 0, 1], [0, 1, 2]],
-    "S": [[-1, -2, -1], [0, 0, 0], [1, 2, 1]],
-    "SW": [[0, -2, -1], [1, 0, -1], [1, 2, 0]],
-    "W": [[1, 0, -1], [2, 0, -2], [1, 0, -1]],
-    "NW": [[2, 1, 0], [1, 0, -1], [0, -1, -2]],
-    "N": [[1, 2, 1], [0, 0, 0], [-1, -2, -1]],
-    "NE": [[0, 1, 2], [-1, 0, 1], [-2, -1, 0]],
-}
-
-median_blur_filters = {
-    "3x3": 3,
-    "5x5": 5,
-    "7x7": 7,
-    "9x9": 9,
-}
-
-prewitt_filters = {
-    "x": [[-1, 0, 1], [-1, 0, 1], [-1, 0, 1]],
-    "y": [[-1, -1, -1], [0, 0, 0], [1, 1, 1]],
-}
